Use #!/bin/bash instead of #!/bin/sh
roundup.sh uses the ERR trap to catch failed assumptions in
testcases.

In http://www.opengroup.org/onlinepubs/009695399/utilities/trap.html
POSIX says that 'trap ERR' is "allowable as an extension" but does
not mandate it, e.g. bash implements it, but dash does not.

Setting the shebang to #!/bin/bash guarantees the ERR trap to
be available.