Add `expectfail` operator as a ! replacement which triggers errors
Asserting an non-zero return code of a command using !, e.g.

  ! grep "string"

does not trigger the error trap of bash. Hence, finding "string" in the
example does not result in a failed testcase in roundup.

To express those assertions, an `expectfail` operator is added which
is equivalent to !, but is handled by bash's error trap:

  expectfail grep "string"