Write stdout/stderr files into $roundup_tmp/... rather then into current dir
The 'capture <cmd>' function captures the output of <cmd>
to the files 'stdout' and 'stderr', which then can be inspected for
expected output.

Currently these files are created inside the current directory.

This patch will change the creation place of the files to be inside
$roundup_tmp.
It will also add two helper functions 'stdout()' and 'stderr()' which will
return the path to $roundup_tmp/stdout and $roundup_tmp/stderr respectively.

NOTE:
Current test-cases that utilize stdout/stderr have to be modified:

Old:
----
 it_tests_tmp_dir_is_not_empty() {
   capture ls /tmp
   test -s stdout
 }

 it_tests_mnt_dir_is_empty() {
   capture ls /mnt
   expectfail test -s stdout
 }

New:
----
 it_tests_tmp_dir_is_not_empty() {
   capture ls /tmp
   test -s $(stdout)
 }

 it_tests_mnt_dir_is_empty() {
   capture ls /mnt
   expectfail test -s $(stdout)
 }