Extend roundup.sh with init() and cleanup()
roundup will call a init() function in each ..._test.sh module,
if present.

This will be called once per module in opposite to the before()
function which is called before each test case

After the last test case in the test module, or if interrupted with
e.g. CTRL+C, the cleanup() function will be called.