Test if all tests in a test plan are actually executed
The number of Tests found in a plan is compared with the number of
executed tests. If there is a mismatch we fail the plan.

This also covers test scripts that have syntax errors since we count
the number of tests via grep. When no tests are executed because sourcing
of the script fails this will create a test count mismatch.